requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=5.4.0
scipy>=1.7.0
scikit-learn>=0.24.0
biopython>=1.79
tqdm>=4.62.0